Rename ./components/basic to ./components. Move back src/c/b/styles to src/styles
