+++ src/components/Lane/LaneFooter.js
@@ -1,10 +1,10 @@
 import React from 'react'
 
-import {LaneFooter} from 'components/basic/styles/Base'
+import {LaneFooter} from 'styles/Base'
 
 import {
   CollapseBtn,
   ExpandBtn,
-} from 'components/basic/styles/Elements'
+} from 'styles/Elements'
 
 export default ({onClick, collapsed}) => <LaneFooter onClick={onClick}>{collapsed ? <ExpandBtn /> : <CollapseBtn />}</LaneFooter>
