+++ src/components/Loader.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import {LoaderDiv, LoadingBar} from './styles/Loader'
+import {LoaderDiv, LoadingBar} from 'styles/Loader'
 
 const Loader = () => (
   <LoaderDiv>
