+++ src/components/NewCardForm.js
@@ -7,8 +7,8 @@ import {
   CardTitle,
   CardWrapper,
   Detail
-} from './styles/Base'
-import {AddButton, CancelButton} from './styles/Elements'
+} from 'styles/Base'
+import {AddButton, CancelButton} from 'styles/Elements'
 import EditableLabel from './widgets/EditableLabel'
 
 class NewCardForm extends Component {
