+++ src/components/widgets/DeleteButton.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import {DeleteWrapper, DelButton} from '../styles/Elements'
+import {DeleteWrapper, DelButton} from 'styles/Elements'
 
 const DeleteButton = props => {
   return (
