+++ src/components/widgets/EditableLabel.js
